# openapi_check.json
{"openapi":"3.1.0","info":{"title":"DolphinID","description":"Ferramenta de identificaÃ§Ã£o automÃ¡tica de botos via foto-identificaÃ§Ã£o de nadadeira dorsal","version":"1.0.0"},"paths":{"/api/sessions":{"get":{"tags":["Sessions"],"summary":"List Sessions","description":"List all processing sessions, newest first.","operationId":"list_sessions_api_sessions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SessionResponse"},"type":"array","title":"Response List Sessions Api Sessions Get"}}}}}},"post":{"tags":["Sessions"],"summary":"Create Session","description":"Create a new processing session and start the pipeline.","operationId":"create_session_api_sessions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions/{session_id}":{"get":{"tags":["Sessions"],"summary":"Get Session Progress","description":"Get detailed progress for a session.","operationId":"get_session_progress_api_sessions__session_id__get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionProgressResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions/{session_id}/results":{"get":{"tags":["Sessions"],"summary":"Get Session Results","description":"Get all results for a session with optional filtering.","operationId":"get_session_results_api_sessions__session_id__results_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"min_confidence","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Confidence"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/results/{result_id}/confirm":{"post":{"tags":["Results"],"summary":"Confirm Result","description":"Confirm or correct the identification of a result.","operationId":"confirm_result_api_results__result_id__confirm_post","parameters":[{"name":"result_id","in":"path","required":true,"schema":{"type":"integer","title":"Result Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/results/{result_id}/crop":{"get":{"tags":["Results"],"summary":"Serve Crop","description":"Serve the crop image for a result.","operationId":"serve_crop_api_results__result_id__crop_get","parameters":[{"name":"result_id","in":"path","required":true,"schema":{"type":"integer","title":"Result Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/results/{result_id}/original":{"get":{"tags":["Results"],"summary":"Serve Original","description":"Serve the original image for a result.","operationId":"serve_original_api_results__result_id__original_get","parameters":[{"name":"result_id","in":"path","required":true,"schema":{"type":"integer","title":"Result Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/gallery/individuals":{"get":{"tags":["Gallery"],"summary":"List Individuals","description":"List all known individuals with their photo counts.","operationId":"list_individuals_api_gallery_individuals_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/gallery/individuals/{label}":{"get":{"tags":["Gallery"],"summary":"Get Individual Detail","description":"Get detailed info for a specific individual including all gallery images.","operationId":"get_individual_detail_api_gallery_individuals__label__get","parameters":[{"name":"label","in":"path","required":true,"schema":{"type":"string","title":"Label"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/gallery/individuals/{label}/image/{index}":{"get":{"tags":["Gallery"],"summary":"Serve Gallery Image","description":"Serve a gallery image for a specific individual.","operationId":"serve_gallery_image_api_gallery_individuals__label__image__index__get","parameters":[{"name":"label","in":"path","required":true,"schema":{"type":"string","title":"Label"}},{"name":"index","in":"path","required":true,"schema":{"type":"integer","title":"Index"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/gallery/embedding-map":{"get":{"tags":["Gallery"],"summary":"Get Embedding Map","description":"Get a 2D projection (t-SNE) of all gallery embeddings for visualization.\n\nThis is computed on-demand and cached in memory.","operationId":"get_embedding_map_api_gallery_embedding_map_get","parameters":[{"name":"max_points","in":"query","required":false,"schema":{"type":"integer","maximum":2000,"default":500,"title":"Max Points"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/export/{session_id}/csv":{"get":{"tags":["Export"],"summary":"Export Csv","description":"Export session results as a CSV file.","operationId":"export_csv_api_export__session_id__csv_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/export/{session_id}/report":{"get":{"tags":["Export"],"summary":"Export Report","description":"Generate an HTML report for a processing session.","operationId":"export_report_api_export__session_id__report_get","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Serve Index","description":"Serve the main frontend page.","operationId":"serve_index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","description":"Health check.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/config":{"get":{"summary":"Get Config","description":"Return current configuration (for frontend display).","operationId":"get_config_api_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ConfirmRequest":{"properties":{"confirmed_id":{"type":"string","title":"Confirmed Id"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["confirmed_id"],"title":"ConfirmRequest","description":"Request to confirm or correct a result."},"CreateSessionRequest":{"properties":{"source_dir":{"type":"string","title":"Source Dir"},"year":{"type":"integer","title":"Year"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["source_dir","year"],"title":"CreateSessionRequest","description":"Request body for creating a new processing session."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SessionProgressResponse":{"properties":{"id":{"type":"integer","title":"Id"},"status":{"type":"string","title":"Status"},"total_images":{"type":"integer","title":"Total Images"},"processed_images":{"type":"integer","title":"Processed Images"},"failed_images":{"type":"integer","title":"Failed Images"},"progress_percent":{"type":"number","title":"Progress Percent"},"status_breakdown":{"additionalProperties":{"type":"integer"},"type":"object","title":"Status Breakdown"}},"type":"object","required":["id","status","total_images","processed_images","failed_images","progress_percent","status_breakdown"],"title":"SessionProgressResponse","description":"Detailed progress of a processing session."},"SessionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"year":{"type":"integer","title":"Year"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"source_dir":{"type":"string","title":"Source Dir"},"status":{"type":"string","title":"Status"},"total_images":{"type":"integer","title":"Total Images"},"processed_images":{"type":"integer","title":"Processed Images"},"failed_images":{"type":"integer","title":"Failed Images"},"created_at":{"type":"string","title":"Created At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","name","year","notes","source_dir","status","total_images","processed_images","failed_images","created_at","completed_at"],"title":"SessionResponse","description":"Session data returned to the frontend."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}